refactor & extract generate-test
